pyproject.toml
[tool.poetry]
name = "pytest-serverless"
version = "0.3.0"
description = "Automatically mocks resources from serverless.yml file."
authors = ["Daniel Ancuta <whisller@gmail.com>"]
readme = "README.md"
include = ["LICENCE", "README.md"]
homepage = "https://github.com/whisller/pytest-serverless"
repository = "https://github.com/whisller/pytest-serverless"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3 :: Only"
]

[tool.poetry.dependencies]
python = "^3.6"
pytest = "^4.5"
pyyaml = "^5.1"
moto = "^1.3"
boto3 = "^1.9"
python-box = "^3.4"

[tool.poetry.dev-dependencies]
black = "=19.3b0"

[tool.poetry.plugins."pytest11"]
"serverless" = "pytest_serverless"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
